# perf(serialize): faster serialization of object entries