@zsilbi 2025-03-04
I realized that this can be simplified if `serializeObjectEntries()` accepts only tuples, so we don't even need the `isArray()` check.