@zsilbi 2025-03-04
New benchmarks:

```scala
 ✓ test/benchmarks.bench.ts > benchmarks > serialize > count:1, size:small 1455ms
     name                         hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · serialize @ main     722,403.10  0.0012  0.5816  0.0014  0.0013  0.0022  0.0026  0.0165  ±0.39%   361202
   · serialize @ current  756,127.23  0.0012  0.3027  0.0013  0.0013  0.0017  0.0021  0.0051  ±0.26%   378064   fastest

 ✓ test/benchmarks.bench.ts > benchmarks > serialize > count:256, size:small 1208ms
     name                       hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · serialize @ main     6,415.09  0.1333  2.9150  0.1559  0.1461  0.3426  0.3964  0.7024  ±1.58%     3208
   · serialize @ current  7,221.52  0.1283  0.3535  0.1385  0.1349  0.2272  0.2325  0.2620  ±0.44%     3611   fastest

 BENCH  Summary

  serialize @ current - test/benchmarks.bench.ts > benchmarks > serialize > count:1, size:small
    1.05x faster than serialize @ main

  serialize @ current - test/benchmarks.bench.ts > benchmarks > serialize > count:256, size:small
    1.13x faster than serialize @ main

clk: ~5.30 GHz
cpu: AMD Ryzen 9 7950X 16-Core Processor
runtime: bun 1.2.4 (x64-linux)

benchmark                   avg (min … max) p75 / p99    (min … top 1%)
------------------------------------------- -------------------------------
• count:1, size:small
------------------------------------------- -------------------------------
serialize @ main             460.10 ns/iter 449.66 ns  █
                    (429.27 ns … 809.25 ns) 714.71 ns  █
                    (  0.00  b … 660.00  b)  12.46  b ▅█▅▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

serialize                    396.05 ns/iter 380.63 ns  █
                    (362.08 ns … 785.06 ns) 653.08 ns  █
                    (  0.00  b … 990.00  b)  29.57  b ▅█▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

summary
  serialize
   1.16x faster than serialize @ main

• count:256, size:small
------------------------------------------- -------------------------------
serialize @ main             479.48 ns/iter 460.29 ns █
                    (444.69 ns … 858.86 ns) 836.80 ns ██
                    (  0.00  b …   1.48 kb)  41.67  b ██▂▁▁▁▂▂▂▁▁▁▂▁▁▁▁▁▁▁▁

serialize                    397.00 ns/iter 383.04 ns  █
                    (368.61 ns … 637.99 ns) 606.37 ns  █
                    (  0.00  b … 924.00  b)  26.80  b ▆█▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂

summary
  serialize
   1.21x faster than serialize @ main
```