@zsilbi 2025-03-04
`compare()` is now only used for Sets, so I inlined it directly into the sorting function..

```scala
 ✓ test/benchmarks.bench.ts > benchmarks > serialize > count:128, size:large 1217ms
     name                       hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · serialize @ main     1,215.50  0.6826  2.6326  0.8227  0.7539  1.8924  2.1144  2.6326  ±2.47%      608
   · serialize @ current  1,466.28  0.6450  1.2528  0.6820  0.6768  0.9048  1.0988  1.2528  ±0.61%      734   fastest

 BENCH  Summary

  serialize @ current - test/benchmarks.bench.ts > benchmarks > serialize > count:128, size:large
    1.21x faster than serialize @ main
```