@zsilbi 2025-03-04
I'll close this because I found an issue with Maps.
I'll open a new PR after https://github.com/unjs/ohash/pull/139 is merged.